+++ docs/misc/changelog.md
@@ -47,6 +47,7 @@
 ### New Features:
 
 - Added official support for Python 3.13
+- Use MacOS Metal "MPS" device when available
 
 ### Bug Fixes: